Changes for 2-3 components
The option to work with 2 and 3 components was not working properly, just deleting data when one was not inside thresholds. This is resolved but needs to be checked, possibly with the AVN dataset.

I have also inserted a "data warning" function. It tells you how many data are below threshold and saves the path of the corresponding file in Murat.mat. The corresponding names are Murat.data.problem**

Co-Authored-By: SimonaGabrielli <77244599+SimonaGabrielli@users.noreply.github.com>